requirements.txt
requests~=2.32.3
psycopg2-binary==2.9.6
python-dotenv==1.0.0